refactor(utils): 将验证逻辑提取到独立的validators模块
将配置验证、网盘类型验证等逻辑从config.ts提取到新的validators.ts模块
保持原有功能不变，提高代码组织性和可维护性
